moved platform inside module
